rename ce to calibrated_explanations
as this is needed for publication at PyPI, see:

https://github.com/Moffran/calibrated_explanations/issues/3

I initially tried to name the package ce on PyPI, but that name is reserved by
PyPI and I get an error message -- I think it is too short. So I guess it makes
sense to go with a more descriptive name "calibrated explanations".

PEP 423 details the rules for packages published at PyPI. The main guideline is
that the module name should match the package name. Following the guidelines
there (which I won't repeat here for brevity), my suggestion is:

    calibrated_explanations as the module name
    calibrated-explanations as the package name